+++ src/consts.ts
@@ -1,3 +1,5 @@
+import { Cityes } from './types/offer';
+
 export enum AppRoute {
   Main = '/',
   Login = '/login',
@@ -41,3 +43,55 @@ export const cityNames = ['Paris', 'Cologne', 'Brussels', 'Amsterdam', 'Hamburg'
 export const URL_MARKER_DEFAULT = 'https://assets.htmlacademy.ru/content/intensive/javascript-1/demo/interactive-map/pin.svg';
 
 export const URL_MARKER_CURRENT = 'https://assets.htmlacademy.ru/content/intensive/javascript-1/demo/interactive-map/main-pin.svg';
+
+
+export const CITYES:Cityes = [
+  {
+    'name': 'Paris',
+    'location': {
+      'latitude': 48.85661,
+      'longitude': 2.351499,
+      'zoom': 13
+    }
+  },
+  {
+    'name': 'Cologne',
+    'location': {
+      'latitude': 50.938361,
+      'longitude': 6.959974,
+      'zoom': 13
+    }
+  },
+  {
+    'name': 'Brussels',
+    'location': {
+      'latitude': 50.846557,
+      'longitude': 4.351697,
+      'zoom': 13
+    }
+  },
+  {
+    'name': 'Amsterdam',
+    'location': {
+      'latitude': 52.35514938496378,
+      'longitude': 4.673877537499948,
+      'zoom': 10
+    }
+  },
+  {
+    'name': 'Hamburg',
+    'location': {
+      'latitude': 53.550341,
+      'longitude': 10.000654,
+      'zoom': 13
+    }
+  },
+  {
+    'name': 'Dusseldorf',
+    'location': {
+      'latitude': 51.225402,
+      'longitude': 6.776314,
+      'zoom': 13
+    }
+  },
+];
