--- src/mocks/more-offer.ts
@@ -1,239 +1,0 @@
-import { FullOffers } from '../types/offer';
-
-export const fullOffers:FullOffers = [
-  {
-    'id': '4cad0267-552b-4382-9510-b1669498c13f',
-    'title': 'Canal View Prinsengracht',
-    'description': 'Discover daily local life in city center, friendly neighborhood, clandestine casino, karaoke, old-style artisans, art gallery and artist studio downstairs.',
-    'type': 'hotel',
-    'price': 402,
-    'images': [
-      'https://13.design.pages.academy/static/hotel/15.jpg',
-      'https://13.design.pages.academy/static/hotel/8.jpg',
-      'https://13.design.pages.academy/static/hotel/19.jpg',
-      'https://13.design.pages.academy/static/hotel/10.jpg',
-      'https://13.design.pages.academy/static/hotel/5.jpg',
-      'https://13.design.pages.academy/static/hotel/11.jpg'
-    ],
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.868610000000004,
-      'longitude': 2.342499,
-      'zoom': 16
-    },
-    'goods': [
-      'Washing machine',
-      'Cable TV',
-      'Air conditioning'
-    ],
-    'host': {
-      'isPro': true,
-      'name': 'Angelina',
-      'avatarUrl': 'https://13.design.pages.academy/static/host/avatar-angelina.jpg'
-    },
-    'isPremium': false,
-    'isFavorite': false,
-    'rating': 1.8,
-    'bedrooms': 1,
-    'maxAdults': 2,
-    'reviews': ['b041999e-e959-4452-aea1-06561bcd78e3','91ccc8f7-688a-4d9b-8998-a2f9713c1a6c']
-  },
-  {
-    'id': 'e4d1588c-9372-4d2c-a08b-caaf23f668e6',
-    'title': 'The house among olive ',
-    'description': 'Discover daily local life in city center, friendly neighborhood, clandestine casino, karaoke, old-style artisans, art gallery and artist studio downstairs.',
-    'type': 'hotel',
-    'price': 402,
-    'images': [
-      'https://13.design.pages.academy/static/hotel/15.jpg',
-      'https://13.design.pages.academy/static/hotel/8.jpg',
-      'https://13.design.pages.academy/static/hotel/19.jpg',
-      'https://13.design.pages.academy/static/hotel/10.jpg',
-      'https://13.design.pages.academy/static/hotel/5.jpg',
-      'https://13.design.pages.academy/static/hotel/11.jpg'
-    ],
-    'city': {
-      'name': 'Amsterdam',
-      'location': {
-        'latitude': 52.35514938496378,
-        'longitude': 4.673877537499948,
-        'zoom': 8
-      }},
-    'location': {
-      'latitude': 52.3909553943508,
-      'longitude': 4.85309666406198,
-      'zoom': 16
-    },
-    'goods': [
-      'Washing machine',
-      'Cable TV',
-      'Air conditioning'
-    ],
-    'host': {
-      'isPro': true,
-      'name': 'Angelina',
-      'avatarUrl': 'https://13.design.pages.academy/static/host/avatar-angelina.jpg'
-    },
-    'isPremium': false,
-    'isFavorite': false,
-    'rating': 1.8,
-    'bedrooms': 1,
-    'maxAdults': 2,
-    'reviews': ['b041999e-e959-4452-aea1-06561bcd78e3','91ccc8f7-688a-4d9b-8998-a2f9713c1a6c']
-  },
-  {
-    'id': '4157daa8-b21d-4961-a465-94323176ba12',
-    'title': 'House in countryside',
-    'description': 'I rent out a very sunny and bright apartment only 7 minutes walking distance to the metro station. The apartment has a spacious living room with a kitchen, one bedroom and a bathroom with mit bath. A terrace can be used in summer.',
-    'type': 'house',
-    'price': 368,
-    'images': [
-      'https://13.design.pages.academy/static/hotel/8.jpg',
-      'https://13.design.pages.academy/static/hotel/16.jpg',
-      'https://13.design.pages.academy/static/hotel/3.jpg',
-      'https://13.design.pages.academy/static/hotel/19.jpg',
-      'https://13.design.pages.academy/static/hotel/12.jpg',
-      'https://13.design.pages.academy/static/hotel/15.jpg'
-    ],
-    'city': {
-      'name': 'Amsterdam',
-      'location': {
-        'latitude': 52.35514938496378,
-        'longitude': 4.673877537499948,
-        'zoom': 8
-      }},
-    'location': {
-      'latitude': 52.3609553943508,
-      'longitude': 4.85309666406198,
-      'zoom': 16
-    },
-    'goods': [
-      'Towels',
-      'Washer',
-      'Baby seat',
-      'Breakfast',
-      'Dishwasher',
-      'Cable TV',
-      'Coffee machine',
-      'Washing machine',
-      'Laptop friendly workspace',
-      'Kitchen'
-    ],
-    'host': {
-      'isPro': true,
-      'name': 'Angelina',
-      'avatarUrl': 'https://13.design.pages.academy/static/host/avatar-angelina.jpg'
-    },
-    'isPremium': true,
-    'isFavorite': true,
-    'rating': 4.2,
-    'bedrooms': 3,
-    'maxAdults': 3,
-    'reviews': ['cee8472f-9c2f-4626-821c-1fcd607f7824','aaed51bd-0777-4aa7-966f-583354a24bef']
-  },
-  {
-    'id': '6852faac-d558-4549-9ab8-229255ee863e',
-    'title': 'Amazing and Extremely Central Flat',
-    'description': 'Discover daily local life in city center, friendly neighborhood, clandestine casino, karaoke, old-style artisans, art gallery and artist studio downstairs.',
-    'type': 'room',
-    'price': 258,
-    'images': [
-      'https://13.design.pages.academy/static/hotel/6.jpg',
-      'https://13.design.pages.academy/static/hotel/8.jpg',
-      'https://13.design.pages.academy/static/hotel/11.jpg',
-      'https://13.design.pages.academy/static/hotel/2.jpg',
-      'https://13.design.pages.academy/static/hotel/1.jpg',
-      'https://13.design.pages.academy/static/hotel/5.jpg'
-    ],
-    'city': {
-      'name': 'Amsterdam',
-      'location': {
-        'latitude': 52.35514938496378,
-        'longitude': 4.673877537499948,
-        'zoom': 8
-      }},
-    'location': {
-      'latitude': 52.3909553943508,
-      'longitude': 4.929309666406198,
-      'zoom': 16
-    },
-    'goods': [
-      'Air conditioning',
-      'Heating',
-      'Fridge',
-      'Kitchen',
-      'Baby seat',
-      'Washing machine',
-      'Breakfast',
-      'Washer'
-    ],
-    'host': {
-      'isPro': true,
-      'name': 'Angelina',
-      'avatarUrl': 'https://13.design.pages.academy/static/host/avatar-angelina.jpg'
-    },
-    'isPremium': true,
-    'isFavorite': true,
-    'rating': 4.2,
-    'bedrooms': 1,
-    'maxAdults': 1,
-    'reviews': ['5f17ddab-42f1-4431-aaa2-6c18cdab6daf','9f8d3606-a033-4f43-acdc-668cd3d37426']
-
-  },
-  {
-    'id': 'f6e517e8-d266-48f0-9f61-8d2f2b3e6be6',
-    'title': 'Amazing and Extremely Central Flat',
-    'description': 'I rent out a very sunny and bright apartment only 7 minutes walking distance to the metro station. The apartment has a spacious living room with a kitchen, one bedroom and a bathroom with mit bath. A terrace can be used in summer.',
-    'type': 'hotel',
-    'price': 116,
-    'images': [
-      'https://13.design.pages.academy/static/hotel/7.jpg',
-      'https://13.design.pages.academy/static/hotel/18.jpg',
-      'https://13.design.pages.academy/static/hotel/6.jpg',
-      'https://13.design.pages.academy/static/hotel/13.jpg',
-      'https://13.design.pages.academy/static/hotel/1.jpg',
-      'https://13.design.pages.academy/static/hotel/16.jpg'
-    ],
-    'city': {
-      'name': 'Amsterdam',
-      'location': {
-        'latitude': 52.35514938496378,
-        'longitude': 4.673877537499948,
-        'zoom': 8
-      }},
-    'location': {
-      'latitude': 52.3809553943508,
-      'longitude': 4.939309666406198,
-      'zoom': 16
-    },
-    'goods': [
-      'Washing machine',
-      'Dishwasher',
-      'Breakfast',
-      'Wi-Fi',
-      'Fridge',
-      'Heating',
-      'Laptop friendly workspace',
-      'Towels',
-      'Coffee machine'
-    ],
-    'host': {
-      'isPro': true,
-      'name': 'Angelina',
-      'avatarUrl': 'https://13.design.pages.academy/static/host/avatar-angelina.jpg'
-    },
-    'isPremium': false,
-    'isFavorite': false,
-    'rating': 1.5,
-    'bedrooms': 5,
-    'maxAdults': 4,
-    'reviews': ['5d7af522-34c1-45c2-87eb-6060681bb2c2']
-  },
-];
