--- src/mocks/reviews.ts
@@ -1,82 +1,0 @@
-import { Reviews } from '../types/reviews';
-
-export const avalibleReviews: Reviews = [
-  {
-    'id': 'b041999e-e959-4452-aea1-06561bcd78e3',
-    'comment': 'I stayed here for one night and it was an unpleasant experience.',
-    'date': '2023-06-26T21:00:00.436Z',
-    'rating': 1,
-    'user': {
-      'name': 'Jack',
-      'avatarUrl': 'https://13.design.pages.academy/static/avatar/5.jpg',
-      'isPro': true
-    }
-  },
-  {
-    'id': '91ccc8f7-688a-4d9b-8998-a2f9713c1a6c',
-    'comment': 'We loved it so much, the house, the veiw, the location just great.. Highly reccomend :)',
-    'date': '2023-06-26T21:00:00.436Z',
-    'rating': 2,
-    'user': {
-      'name': 'Kendall',
-      'avatarUrl': 'https://13.design.pages.academy/static/avatar/10.jpg',
-      'isPro': false
-    }
-  },
-  {
-    'id': 'cee8472f-9c2f-4626-821c-1fcd607f7824',
-    'comment': 'This villa is perfect in every way: the view on mountains and waterfalls, the hot tub and the villa itself. The evening here became a great continuation of our journeys over country.',
-    'date': '2023-06-28T21:00:00.436Z',
-    'rating': 4,
-    'user': {
-      'name': 'Christina',
-      'avatarUrl': 'https://13.design.pages.academy/static/avatar/10.jpg',
-      'isPro': true
-    }
-  },
-  {
-    'id': 'aaed51bd-0777-4aa7-966f-583354a24bef',
-    'comment': 'I stayed here for one night and it was an unpleasant experience.',
-    'date': '2023-06-28T21:00:00.436Z',
-    'rating': 4,
-    'user': {
-      'name': 'Zak',
-      'avatarUrl': 'https://13.design.pages.academy/static/avatar/8.jpg',
-      'isPro': false
-    }
-  },
-  {
-    'id': '5f17ddab-42f1-4431-aaa2-6c18cdab6daf',
-    'comment': 'The room was spacious and clean. The pool looked nothing like the photos and desparately needs a clean. The sauna and spa were closed for lunar new year holiday.',
-    'date': '2023-06-27T21:00:00.436Z',
-    'rating': 2,
-    'user': {
-      'name': 'Sophie',
-      'avatarUrl': 'https://13.design.pages.academy/static/avatar/1.jpg',
-      'isPro': false
-    }
-  },
-  {
-    'id': '9f8d3606-a033-4f43-acdc-668cd3d37426',
-    'comment': 'The house is very good, very happy, hygienic and simple living conditions around it are also very good. I hope to have the opportunity to come back. Thank you.',
-    'date': '2023-06-24T21:00:00.436Z',
-    'rating': 4,
-    'user': {
-      'name': 'Mollie',
-      'avatarUrl': 'https://13.design.pages.academy/static/avatar/1.jpg',
-      'isPro': true
-    }
-  },
-  {
-    'id': '5d7af522-34c1-45c2-87eb-6060681bb2c2',
-    'comment': 'We loved it so much, the house, the veiw, the location just great.. Highly reccomend :)',
-    'date': '2023-06-26T21:00:00.436Z',
-    'rating': 4,
-    'user': {
-      'name': 'Max',
-      'avatarUrl': 'https://13.design.pages.academy/static/avatar/9.jpg',
-      'isPro': true
-    }
-  }
-];
-
